+++ wradlib/util.py
@@ -639,33 +639,5 @@ def gradient_from_smoothed(x, n=5):
     return gradient_along_axis(medfilt_along_axis(x, n)).astype("f4")
 
 
-def grid_to_polyvert(grid, ravel=False):
-    """Get polygonal vertices from rectangular grid coordinates.
-
-    Parameters
-    ----------
-    grid : numpy array
-        grid edge coordinates
-
-    Returns
-    -------
-    polyvert : :class:`numpy:numpy.ndarray`
-        A 3-d array of polygon vertices with shape (..., 5, 2).
-
-    """
-
-    v1 = grid[:-1, :-1]
-    v2 = grid[:-1, 1:]
-    v3 = grid[1:, 1:]
-    v4 = grid[1:, :-1]
-
-    polyvert = np.stack((v1, v2, v3, v4, v1), axis=-2)
-
-    if ravel:
-        polyvert = polyvert.reshape((-1, 5, 2))
-
-    return(polyvert)
-
-
 if __name__ == '__main__':
     print('wradlib: Calling module <util> as main...')
